% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleanovas.R
\name{ANOVAsAsTable}
\alias{ANOVAsAsTable}
\title{\code{ANOVAsAsTable}
Converts a list of ANOVAs into a format that can be prettily formatted.}
\usage{
ANOVAsAsTable(x)
}
\arguments{
\item{x}{The list of ANOVAs.}
}
\description{
\code{ANOVAsAsTable}
Converts a list of ANOVAs into a format that can be prettily formatted.
}

