% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableofdifferences.R
\name{autoFontColor}
\alias{autoFontColor}
\title{Use black or white for good contrast against colors}
\usage{
autoFontColor(colors)
}
\arguments{
\item{colors}{vector of colors which will be the background color of the}
}
\description{
Use black or white for good contrast against colors
}
