% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparemeans.R
\name{CompareMeans}
\alias{CompareMeans}
\title{\code{CompareMeans}
Performs statistical tests comparing means of an outcome variable.}
\usage{
CompareMeans(outcome, columns, Rows = NULL, compare = "Columns",
  correction = "False Discovery Rate", show.labels = TRUE,
  output = "Pretty", outcome.name = NULL, ...)
}
\arguments{
\item{outcome}{The outcome variable.}

\item{columns}{The factor representing the columns.}

\item{compare}{One of \code{"Rows", "Columns", "Columns pairwise"}, or \code{"All"}}

\item{correction}{The multiple comparison adjustment method: \code{"None",
"False Discovery Rate", "Benjamini & Yekutieli", "Bonferroni",
"Free Combinations"} (Westfall et al. 1999), \code{"Hochberg", "Holm",
"Hommel", "Single-step"} (Bretz et al. 2010) \code{"Shaffer"}, and \code{"Westfall"}.}

\item{show.labels}{Shows the variable labels, as opposed to the labels, in the outputs, where a
variables label is an attribute (e.g., attr(foo, "label")).}

\item{output}{One of \code{"Pretty"} or \code{"Table"}.}

\item{rows}{The list of factors representing the rows.}

\item{subset}{An optional vector specifying a subset of observations to be
used in the fitting process, or, the name of a variable in \code{data}. It
may not be an expression. \code{subset} may not}

\item{weights}{An optional vector of sampling weights, or, the name or, the
name of a variable in \code{data}. It may not be an expression.}

\item{type}{Defaults to \code{"linear"}. Other types are: \code{"Poisson"},
\code{"Quasi-Poisson"}, \code{"Binary Logit"}, \code{"NBD"}, and
\code{"Ordered Logit"}.}

\item{missing}{How missing data is to be treated in the regression. Options:
\code{"Error if missing data"},
\code{"Exclude cases with missing data"},
\code{"Use partial data (pairwise correlations)"},
\code{"Imputation (replace missing values with estimates)"}, and
\code{"Multiple imputation"}.}
}
\description{
\code{CompareMeans}
Performs statistical tests comparing means of an outcome variable.
}
\references{
Bretz,Frank, Torsten Hothorn and Peter Westfall (2010), Multiple Comparisons Using R, CRC Press, Boca Raton.
Benjamini, Y., and Hochberg, Y. (1995). Controlling the false discovery rate: a practical and powerful approach to multiple testing. Journal of the Royal Statistical Society Series B 57, 289–300.
Benjamini, Y., and Yekutieli, D. (2001). The control of the false discovery rate in multiple testing under dependency. Annals of Statistics 29, 1165–1188.
Holm, S. (1979). A simple sequentially rejective multiple test procedure. Scandinavian Journal of Statistics 6, 65–70.
Hochberg, Y. (1988). A sharper Bonferroni procedure for multiple tests of significance. Biometrika 75, 800–803.
Hommel, G. (1988). A stagewise rejective multiple test procedure based on a modified Bonferroni test. Biometrika 75, 383–386.
Hothorn, Torsten, Frank Bretz and Peter Westfall (2008), Simultaneous Inference in General Parametric Models. Biometrical Journal, 50(3), 346–363.
Shaffer, Juliet P. (1986), Modified sequentially rejective multiple test procedures. Journal of the American Statistical Association, 81, 826–831.
Shaffer, Juliet P. (1995). Multiple hypothesis testing. Annual Review of Psychology 46, 561–576.
Sarkar, S. (1998). Some probability inequalities for ordered MTP2 random variables: a proof of Simes conjecture. Annals of Statistics 26, 494–504.
Sarkar, S., and Chang, C. K. (1997). Simes' method for multiple hypothesis testing with positively dependent test statistics. Journal of the American Statistical Association 92, 1601–1608.
Peter H. Westfall (1997), Multiple testing of general contrasts using logical constraints and correlations. Journal of the American Statistical Association, 92, 299–306.
P. H. Westfall, R. D. Tobias, D. Rom, R. D. Wolfinger, Y. Hochberg (1999). Multiple Comparisons and Multiple Tests Using the SAS System. Cary, NC: SAS Institute Inc.
Wright, S. P. (1992). Adjusted P-values for simultaneous inference. Biometrics 48, 1005–1013.
}

