% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparemeans.R
\name{CompareMultipleMeans}
\alias{CompareMultipleMeans}
\title{\code{CompareMultipleMeans}
Compares means of multiple variables.}
\usage{
CompareMultipleMeans(variables, groups, title = "", subtitle = "",
  footer = "", weights = NULL, ...)
}
\arguments{
\item{outcome}{The outcome variable.}
}
\description{
\code{CompareMultipleMeans}
Compares means of multiple variables.
}

