% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleanovas.R
\name{FormattableANOVAs}
\alias{FormattableANOVAs}
\title{FormattableANOVAs}
\usage{
FormattableANOVAs(anovas, title, subtitle, footer)
}
\arguments{
\item{anovas}{List of OneWayANOVA objects.}

\item{title}{String showing the title of the table.}

\item{subtitle}{String showing the subtitle of the table.}

\item{footer}{String showing the title of the table.}
}
\description{
FormattableANOVAs
}

