% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onewayanova.R
\name{FTest}
\alias{FTest}
\title{\code{FTest}}
\usage{
FTest(fit)
}
\arguments{
\item{fit}{An object created by \link{FitRegression}.}
}
\description{
An F-Test from a \code{lm} or linear \code{svyglm} object.
}

