% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparemeans.R
\name{MeansTables}
\alias{MeansTables}
\title{\code{CompareMeans}
Performs statistical tests comparing means of an outcome variable.}
\usage{
MeansTables(x)
}
\arguments{
\item{outcome}{The outcome variable.}
}
\description{
\code{CompareMeans}
Performs statistical tests comparing means of an outcome variable.
}

