% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleanovas.R
\name{MultipleANOVAs}
\alias{MultipleANOVAs}
\title{\code{MultipleANOVAs}}
\usage{
MultipleANOVAs(
  dependents,
  independent,
  subset = NULL,
  weights = NULL,
  compare = "To mean",
  correction = "Table FDR",
  robust.se = FALSE,
  alternative = "Two-sided",
  missing = "Exclude cases with missing data",
  show.labels = FALSE,
  seed = 1223,
  p.cutoff = 0.05,
  data = NULL,
  ...
)
}
\arguments{
\item{dependents}{The outcome variables.}

\item{independent}{The factor representing the groups.}

\item{subset}{An optional vector specifying a subset of observations to be
used in the fitting process, or, the name of a variable in \code{data}. It
may not be an expression. \code{subset} may not}

\item{weights}{An optional vector of sampling weights, or, the name or, the
name of a variable in \code{data}. It may not be an expression.}

\item{compare}{One of \code{"To mean", "Pairwise", "To first"} (which implement's Dunnett's C, when
combined with 'correction' == 'Tukey Range'), or \code{"All"}}

\item{correction}{The multiple comparison adjustment method: \code{"Table FDR", "Tukey Range", "None",
"False Discovery Rate", "Benjamini & Yekutieli", "Bonferroni",
"Free Combinations"}, \code{"Hochberg", "Holm",
"Hommel", "Single-step"} \code{"Shaffer"}, and \code{"Westfall"}.}

\item{robust.se}{Computes standard errors that are robust to violations of
the assumption of constant variance. This parameter is ignored
if weights are applied (as weights already employ a sandwich estimator).}

\item{alternative}{The alternative hypothesis: "Two sided", "Greater", or "Less". The main application of this is when
Compare us set 'To first' (e.g., if testing a new product, where the purpose is to work out of the new product is superior
to an existing product, "Greater" would be chosen).}

\item{missing}{How missing data is to be treated in the ANOVA. Options:
\code{"Error if missing data"}.
\code{"Exclude cases with missing data"}, and
\code{"Imputation (replace missing values with estimates)"}.}

\item{show.labels}{Shows the variable labels, as opposed to the labels, in the outputs, where a
variables label is an attribute (e.g., attr(foo, "label")).}

\item{seed}{The random number seed used when evaluating the multivariate t-distribution.}

\item{p.cutoff}{The alpha level to be used in testing.}

\item{data}{An optional \code{data.frame} that can be used as alternative to specifying
\code{dependent}, \code{independent}, and \code{weights} for the case when the independent
variables are deferring lengths. Must have columns
\code{"dependent", "independent", "weights", and "dependent.names"}. The \code{"dependent.names"}
column contains the character name of the dependent variable for each observation. The \code{dependent},
\code{independent}, \code{weights}, and \code{subset} arguments are all ignored if \code{data}
is supplied.}

\item{...}{Other parameters to be passed to \code{OneWayANOVA}.}
}
\description{
Computes multiple ANOVAs.
}
\details{
Conducts multiple \code{OneWayANOVA}s, and puts them in a list. If \code{correction} is
 \code{"Table FDR"}, the false discovery rate correction is applied across the entire table. All
other corrections are performed within rows. Additional detail about the other parameters can be found in \code{OneWayANOVA}.
}
