% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplemeans.R
\name{MultipleMeans}
\alias{MultipleMeans}
\title{\code{MultipleMeans}}
\usage{
MultipleMeans(
  outcomes,
  predictor,
  subset = NULL,
  weights = NULL,
  correction = "Table FDR",
  robust.se = FALSE,
  missing = "Exclude cases with missing data",
  show.labels = FALSE,
  seed = 1223,
  p.cutoff = 0.05,
  title = "",
  subtitle = "",
  footer = "",
  return.data.frame = FALSE
)
}
\arguments{
\item{outcomes}{The outcome variables.}

\item{predictor}{The factor representing the groups.}

\item{subset}{An optional vector specifying a subset of observations to be
used in the fitting process, or, the name of a variable in \code{data}. It
may not be an expression. \code{subset} may not}

\item{weights}{An optional vector of sampling weights, or, the name or, the
name of a variable in \code{data}. It may not be an expression.}

\item{correction}{The multiple comparison adjustment method: \code{"Table FDR", "Tukey Range", "None",
"False Discovery Rate", "Benjamini & Yekutieli", "Bonferroni",
"Free Combinations"}, \code{"Hochberg", "Holm",
"Hommel", "Single-step"} \code{"Shaffer"}, and \code{"Westfall"}.}

\item{robust.se}{Computes standard errors that are robust to violations of
the assumption of constant variance. This parameter is ignored
if weights are applied (as weights already employ a sandwich estimator).}

\item{missing}{How missing data is to be treated in the ANOVA. Options:
\code{"Error if missing data"}.
\code{"Exclude cases with missing data"}, and
\code{"Imputation (replace missing values with estimates)"}.}

\item{show.labels}{Shows the variable labels, as opposed to the labels, in the outputs, where a
variables label is an attribute (e.g., attr(foo, "label")).}

\item{seed}{The random number seed used when evaluating the multivariate t-distribution.}

\item{p.cutoff}{The alpha level to be used in testing.}

\item{title}{The title to appear in the output.}

\item{subtitle}{The footer to appear in the output.}

\item{footer}{The footer to appear in the output.}

\item{return.data.frame}{Whether to return a data frame instead of a
formattable widget.}
}
\description{
Compares means of multiple outcomes on a categorical predictor.
}
\details{
Computes multiple ANOVAs.

Conducts multiple \code{OneWayANOVA}s, and puts them in a list. If \code{correction} is
 \code{"Table FDR"}, the false discovery rate correction is applied across the entire table. All
other corrections are performed within rows. Additional detail about the other parameters can be found in \code{OneWayANOVA}.
}
