% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onewaymanova.R
\name{OneWayMANOVA}
\alias{OneWayMANOVA}
\title{\code{OneWayMANOVA}}
\usage{
OneWayMANOVA(outcomes, predictor, subset = NULL, weights = NULL,
  correction = "Tukey Range", robust.se = FALSE,
  missing = "Exclude cases with missing data", show.labels = FALSE,
  seed = 1223, p.cutoff = 0.05, binary = FALSE, ...)
}
\arguments{
\item{outcomes}{The outcome variables.}

\item{predictor}{The factor representing the groups.}

\item{subset}{An optional vector specifying a subset of observations to be
used in the fitting process, or, the name of a variable in \code{data}. It
may not be an expression. \code{subset} may not}

\item{weights}{An optional vector of sampling weights, or, the name or, the
name of a variable in \code{data}. It may not be an expression.}

\item{correction}{The multiple comparison adjustment method: \code{"Tukey Range", "None",
"False Discovery Rate", "Benjamini & Yekutieli", "Bonferroni",
"Free Combinations"}, \code{"Hochberg", "Holm",
"Hommel", "Single-step"} \code{"Shaffer"}, and \code{"Westfall"}.}

\item{robust.se}{Computes standard errors that are robust to violations of
the assumption of constant variance. This parameter is ignored
if weights are applied (as weights already employ a sandwich estimator).}

\item{missing}{How missing data is to be treated in the ANOVA. Options:
\code{"Error if missing data"}, and
\code{"Exclude cases with missing data"}.}

\item{show.labels}{Shows the variable labels, as opposed to the labels, in the outputs, where a
variables label is an attribute (e.g., attr(foo, "label")).}

\item{seed}{The random number seed used when evaluating the multivariate t-distribution.}

\item{p.cutoff}{The alpha level to be used in testing.}

\item{binary}{Automatically converts non-ordered factors to dummy-coded (binary indicator) variables.}

\item{...}{Other parameters to be passed to \code{OneWayANOVA}.}
}
\description{
Computes a one-way analysis of variance with post hoc tests.
}
\details{
Where sampling weights are provided, a sample is constructed via bootstrapping, and this
sample is used in the testing for the MANOVA (but no for the individual ANOVAs or post hoc tests).
See \code{OneWayANOVA} for more about the other parameters.

Tests are two-sided, comparing to the Grand Mean (i.e., "To mean").

Additional detail about the other parameters can be found in \code{OneWayANOVA}.
}

