% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onewayanova.R
\name{politeWeightedFTest}
\alias{politeWeightedFTest}
\title{politeWeightedFTest}
\usage{
politeWeightedFTest(fit)
}
\arguments{
\item{fit}{A FitRegression object.}
}
\description{
Computes \link{regTermTest}, but fails savely when an error occurs.
}

