% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparemeans.R
\name{print.CompareMultipleMeans}
\alias{print.CompareMultipleMeans}
\title{\code{print.CompareMultipleMeans}}
\usage{
\method{print}{CompareMultipleMeans}(x, ...)
}
\arguments{
\item{x}{An object of class \code{CompareMultipleMeans}}

\item{...}{Arguments for \code{PrintFormattableANOVAs}.}
}
\description{
\code{print.CompareMultipleMeans}
}

