% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onewaymanova.R
\name{print.OneWayMANOVA}
\alias{print.OneWayMANOVA}
\title{print.OneWayMANOVA}
\usage{
\method{print}{OneWayMANOVA}(x, ...)
}
\arguments{
\item{x}{A OneWayMANOVA object.}

\item{...}{Other arguments.}
}
\description{
Returns the OneWayMANOVA as a pretty table.
}
