% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onewaymanova.R
\name{PrintFormattableANOVAs}
\alias{PrintFormattableANOVAs}
\title{PrintFormattableANOVAs}
\usage{
PrintFormattableANOVAs(anovas, title, subtitle, footer)
}
\arguments{
\item{anovas}{List of OneWayANOVA objects.}

\item{title}{String showing the title of the table.}

\item{subtitle}{String showing the subtitle of the table.}

\item{footer}{String showing the title of the table.}
}
\description{
PrintFormattableANOVAs
}

