% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableofmeans.R
\name{TableOfMeans}
\alias{TableOfMeans}
\title{\code{TableOfMeans}}
\usage{
TableOfMeans(outcome, row, column, correction = "Table FDR",
  robust.se = FALSE, weights = NULL, show.labels = TRUE, ...)
}
\arguments{
\item{outcome}{The outcome variable.}

\item{row}{The predictor to show in the rows. Used to filter the data.}

\item{column}{The predictor to show in the columns.}

\item{robust.se}{Computes standard errors that are robust to violations of
the assumption of constant variance. This parameter is ignored
if weights are applied (as weights already employ a sandwich estimator).}

\item{show.labels}{Show labels instead of variable names}

\item{...}{Parameters to pass to \code{\link{OneWayANOVA}}.}
}
\description{
Compares mean of one variable by two factors.
}

