% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparemeans.R
\name{CompareMultipleMeans}
\alias{CompareMultipleMeans}
\title{\code{CompareMultipleMeans}}
\usage{
CompareMultipleMeans(outcomes, predictor, title = "", subtitle = "",
  footer = "", weights = NULL, show.labels = TRUE, ...)
}
\arguments{
\item{outcomes}{The outcome variables.}

\item{predictor}{The factor indicating group membership.}

\item{title}{Main title for the table.}

\item{subtitle}{Subtitle for the table.}

\item{footer}{Footer of the table.}

\item{weights}{The sampling weights}

\item{show.labels}{Show labels instead of variable names}

\item{...}{Parameters to pass to \code{\link{OneWayANOVA}}.}
}
\description{
Compares means of multiple variables.
}

