% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onewaymanova.R
\name{OneWayMANOVA}
\alias{OneWayMANOVA}
\title{\code{OneWayMANOVA}}
\usage{
OneWayMANOVA(outcomes, predictor, subset = NULL, weights = NULL,
  robust.se = FALSE, missing = "Exclude cases with missing data",
  show.labels = FALSE, seed = 1223, p.cutoff = 0.05,
  binary = FALSE, pillai = FALSE, fdr = TRUE, return.all = FALSE,
  ...)
}
\arguments{
\item{outcomes}{The outcome variables.}

\item{predictor}{The factor representing the groups.}

\item{subset}{An optional vector specifying a subset of observations to be
used in the fitting process, or, the name of a variable in \code{data}. It
may not be an expression. \code{subset} may not}

\item{weights}{An optional vector of sampling weights, or, the name or, the
name of a variable in \code{data}. It may not be an expression.}

\item{robust.se}{Computes standard errors that are robust to violations of
the assumption of constant variance. This parameter is ignored
if weights are applied (as weights already employ a sandwich estimator).}

\item{missing}{How missing data is to be treated in the ANOVA. Options:
  \code{"Error if missing data"}, \code{"Use partial data"}, and
  \code{"Exclude cases with missing data"}. Option \code{"Use partial data"} will
run separate one way ANOVAs with missing data filtered out for each outcome variable
separately.}

\item{show.labels}{Shows the variable labels, as opposed to the labels, in the outputs, where a
variables label is an attribute (e.g., attr(foo, "label")).}

\item{seed}{The random number seed used when evaluating the multivariate t-distribution.}

\item{p.cutoff}{The alpha level to be used in testing.}

\item{binary}{Automatically converts non-ordered factors to dummy-coded (binary indicator) variables.}

\item{pillai}{If \code{TRUE}, Pillai's Trace is computed as the overall MANOVA statistic.}

\item{fdr}{If \code{TRUE}, the False Discovery Rate correction is applied. This is the default.}

\item{return.all}{If \code{TRUE}, returns all the internal computations in the output object. If \code{FALSE},
returns just the information required to print the output.}

\item{...}{Other parameters to be passed to \code{OneWayANOVA}.}
}
\description{
Computes a one-way analysis of variance with post hoc tests.
}
\details{
By default, the overall p-value is computed as the smallest p-value in any cell following application of the
False Discovery Rate correction to the p-values. If the\code{fdr} is set to \code{FALSE}, the correction is not applied, which means
that the overall p-value is the smallest of the uncorrected p-values, and, additionally, the p-values for each row
are from the \link{OneWayANOVA} F-tests.

Tests are two-sided, comparing to the Grand Mean (i.e., "To mean" in \link{OneWayANOVA}).

Additional detail about the other parameters can be found in \code{OneWayANOVA}.
}
