% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentcomparisontable.R
\name{SegmentComparisonTable}
\alias{SegmentComparisonTable}
\title{\code{SegmentComparisonTable}}
\usage{
SegmentComparisonTable(x, group, weights = NULL, subset = TRUE,
  remove.net.or.sum = FALSE, format.numeric.decimals = 1,
  format.percentage.decimals = 0, format.conditional.fill = TRUE,
  format.numeric.fill.colors = "#E99598, #E5C8C4, #A9C0DA, #82A5CB",
  format.percentage.fill.colors = "#E99598, #E5C8C4, #A9C0DA, #82A5CB",
  show.index.values = FALSE, cell.fill = "#FFFFFF",
  font.color = "#2C2C2C", font.size = 10, font.unit = "px",
  row.height = paste0(font.size * 1.5, font.unit),
  font.color.set.if.nonsignificant = TRUE,
  font.color.nonsignificant = "#CCCCCC", font.color.confidence = 0.95,
  font.color.FDRcorrection = FALSE, font.color.nonparametric = FALSE,
  show.question.name = TRUE, col.widths = "100px, 100px",
  row.header.font.weight = "normal", row.span.font.weight = "normal",
  col.header.font.weight = "normal", row.header.pad = 2,
  row.span.pad = row.header.pad, summary.cell.fill = "#EFEFEF",
  row.header.fill = "#DDDDDD", row.span.fill = row.header.fill,
  col.header.fill = row.header.fill, corner.fill = row.header.fill,
  ...)
}
\arguments{
\item{x}{A list of variables or variable sets. These can be numeric or
categorical (PickOne or PickAny). It is expected that the attributes
for "question", "label" and "questiontype" is defined.}

\item{group}{The segmentation variable. Should be a factor.}

\item{weights}{Numeric; An optional vector of sampling weights. 
Should be the same length as \code{group}.}

\item{subset}{An optional vector specifying a subset of observations to be used.}

\item{remove.net.or.sum}{Remove NET or SUM rows for Pick Any or NumberMulti variables.}

\item{format.numeric.decimals}{The number of decimals shown in the output table
for numeric values.}

\item{format.percentage.decimals}{The number of decimals shown in the output table
for percentage values.}

\item{format.conditional.fill}{Whether the fill color of the cells should reflect the
value in the cells.}

\item{format.numeric.fill.colors}{A vector or comma-separated list of 5 colors that
is used when \code{format.conditional.fill}. The colors indicate that 
the standardized number value in the cell is smaller than -0.2, -0.1, 0, 0.1, 0.2.
Numeric variables are standardized by dividing twice the standard deviation
http://www.stat.columbia.edu/~gelman/research/published/standardizing7.pdf}

\item{format.percentage.fill.colors}{A vector or comma-separated list of 5 colors that
is used when \code{format.conditional.fill}. The colors indicate that 
the percentage in the cell is smaller than -20\%, -10\%, 0\%, 10\%, 20\%.}

\item{show.index.values}{Percentage values are shown as a ratio to the percentage
computed on the whole population (i.e. unsegmented).}

\item{cell.fill}{The default background color of the cells in the table.}

\item{font.color}{Default font color of the cells in the table.}

\item{font.size}{Default font size.}

\item{font.unit}{One of "px" or "pt".}

\item{row.height}{Height of one row in the table (includes units).}

\item{font.color.set.if.nonsignificant}{Logical; whether p-values should be computed
to test the significance of the value in the table.}

\item{font.color.nonsignificant}{Font color used to show insignificant values if
\code{font.color.set.if.nonsignificant}.}

\item{font.color.confidence}{The confidence level below which insignificant values
should be shown in \code{font.color.nonsignificant} if
\code{font.color.set.if.nonsignficant}.}

\item{font.color.FDRcorrection}{Logical; whether an FDR correction is applied to deal
with multiple testing.}

\item{font.color.nonparametric}{Logical; whether a non-parametric test should}

\item{show.question.name}{Whether the question name should be shown in the output table.}

\item{col.widths}{Vector or comma-separated list to control column widths.}

\item{row.header.font.weight}{One of "bold" or "normal".}

\item{row.span.font.weight}{One of "bold" or "normal".}

\item{col.header.font.weight}{One of "bold" or "normal".}

\item{row.header.pad}{Numeric; Space between border and text in pixels.}

\item{row.span.pad}{Numeric; Space between border and text in pixels.}

\item{summary.cell.fill}{The background color of the first two rows in
the table which gives a summary of the segmentation variable.}

\item{row.header.fill}{Background color of row header cells.}

\item{row.span.fill}{Background color of row span cells.}

\item{col.header.fill}{Background color of column header cells.}

\item{corner.fill}{Background color of corner cells.}

\item{...}{Other parameters passed to \link[flipFormat]{CreateCustomTable}.}
}
\description{
Creates a html table allowing comparison between segments using
variables or variable sets
}
